+++ actionmailer/lib/action_mailer.rb
@@ -44,4 +44,5 @@ module ActionMailer
   autoload :MailHelper
   autoload :TestCase
   autoload :TestHelper
+  autoload :QueuedMessage
 end