+++ actionmailer/test/base_test.rb
@@ -3,11 +3,13 @@ require 'abstract_unit'
 require 'set'
 
 require 'action_dispatch'
+require 'active_support/queueing'
 require 'active_support/time'
 
 require 'mailers/base_mailer'
 require 'mailers/proc_mailer'
 require 'mailers/asset_mailer'
+require 'mailers/async_mailer'
 
 class BaseTest < ActiveSupport::TestCase
   def teardown
@@ -418,6 +420,17 @@ class BaseTest < ActiveSupport::TestCase
     BaseMailer.deliveries.clear
     BaseMailer.welcome.deliver
     assert_equal(1, BaseMailer.deliveries.length)
+  end
+
+  test "delivering message asynchronously" do
+    AsyncMailer.delivery_method = :test
+    AsyncMailer.deliveries.clear
+
+    AsyncMailer.welcome.deliver
+    assert_equal 0, AsyncMailer.deliveries.length
+
+    AsyncMailer.queue.drain
+    assert_equal 1, AsyncMailer.deliveries.length
   end
 
   test "calling deliver, ActionMailer should yield back to mail to let it call :do_delivery on itself" do